test: product의 리팩토링에 따른 테스트 코드 수정
