+++ src/test/java/umc/stockoneqback/common/ControllerTest.java
@@ -48,7 +48,9 @@ import umc.stockoneqback.global.security.handler.JwtAccessDeniedHandler;
 import umc.stockoneqback.global.security.handler.JwtAuthenticationEntryPoint;
 import umc.stockoneqback.global.security.service.CustomUserDetailsService;
 import umc.stockoneqback.product.controller.ProductApiController;
+import umc.stockoneqback.product.controller.ProductFindApiController;
 import umc.stockoneqback.product.service.ProductFindOthersService;
+import umc.stockoneqback.product.service.ProductFindService;
 import umc.stockoneqback.product.service.ProductService;
 import umc.stockoneqback.reply.controller.ReplyApiController;
 import umc.stockoneqback.reply.service.ReplyFindService;
@@ -77,6 +79,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
         CommentApiController.class,
         ReplyApiController.class,
         ProductApiController.class,
+        ProductFindApiController.class,
         FriendApiController.class,
         BoardLikeApiController.class,
         FriendInformationController.class,
@@ -122,6 +125,9 @@ public abstract class ControllerTest {
 
     @MockBean
     protected ProductService productService;
+
+    @MockBean
+    protected ProductFindService productFindService;
 
     @MockBean
     protected UserFindService userFindService;
