+++ src/test/resources/application-firebase.yml
@@ -1,7 +1,7 @@
 firebase:
   type: ${TYPE}
   project_id: ${PROJECT_ID}
-  private_key_id: ${PRIVATE_KEY_ID}가
+  private_key_id: ${PRIVATE_KEY_ID}
   private_key: ${PRIVATE_KEY}
   client_email: ${CLIENT_EMAIL}
   client_id: ${CLIENT_ID}
